# openapi/g7mediabooster-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "g7mb-api",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "tags": [
          "media"
        ],
        "operationId": "capabilities",
        "responses": {
          "200": {
            "description": "Verified native capabilities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed signing headers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "HMAC authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Native capability snapshot is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CapabilitiesResponse": {
        "type": "object",
        "description": "Runtime media capability response.",
        "required": [
          "image_inputs",
          "image_outputs",
          "mp4_thumbnail",
          "mp4_h264_fallback",
          "native_versions"
        ],
        "properties": {
          "image_inputs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Formats the native runtime can decode in this deployment."
          },
          "image_outputs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Formats the native runtime can encode in this deployment."
          },
          "mp4_h264_fallback": {
            "type": "boolean",
            "description": "Whether the narrow MP4/H.264 OpenH264 fallback is compiled into the sandbox."
          },
          "mp4_thumbnail": {
            "type": "boolean",
            "description": "Whether FFprobe and FFmpeg passed the embedded MP4/H.264 runtime fixture."
          },
          "native_versions": {
            "type": "object",
            "description": "Sanitized version lines from the native tools checked by the sandbox.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Stable error envelope returned to API consumers.",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code."
          },
          "message": {
            "type": "string",
            "description": "User-safe explanation without backend details."
          },
          "request_id": {
            "type": "string",
            "description": "Correlation identifier for support and logs."
          }
        }
      },
...
    }
  },
...
}